Move c.h.internal.cluster.impl.Versions to c.h.internal.cluster.Versions
