+++ hazelcast/src/main/java/com/hazelcast/internal/cluster/Versions.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.cluster.impl;
+package com.hazelcast.internal.cluster;
 
 import com.hazelcast.version.Version;
