+++ hazelcast/src/main/java/com/hazelcast/internal/cluster/impl/MembershipManagerCompat.java
@@ -19,6 +19,7 @@ package com.hazelcast.internal.cluster.impl;
 import com.hazelcast.core.Member;
 import com.hazelcast.instance.MemberImpl;
 import com.hazelcast.instance.Node;
+import com.hazelcast.internal.cluster.Versions;
 import com.hazelcast.internal.cluster.impl.operations.MemberRemoveOperation;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.nio.Address;
